requirements.txt
aenum
aiohttp
discord
python-dotenv
pyyaml
motor